Tests moved to its own folder
